moving layout into components folder
