+++ src/App.jsx
@@ -1,6 +1,6 @@
 import { Route, Routes } from "react-router-dom";
 import Home from "./pages/Home";
-import Layout from "./pages/Layout";
+import Layout from "./components/layout/Layout";
 
 export default function App() {
   return (
