+++ src/components/layout/Layout.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Header from "../components/layouts/header/header";
-import Footer from "../components/footer/footer";
+import Header from "./header/Header";
+import Footer from "./footer/Footer";
 
 const Layout = ({ children }) => {
   return (
