+++ src/components/layout/footer/dividerStack.jsx
