+++ src/components/layout/footer/footerInfoStack.jsx
